+++ src/google_play.rs
@@ -19,6 +19,10 @@ pub async fn download_apps(
     let locale = options.remove("locale").unwrap_or("en_US");
     let timezone = options.remove("timezone").unwrap_or("UTC");
     let device = options.remove("device").unwrap_or("hero2lte");
+    let split_apk = match options.remove("split_apk") {
+        Some(val) if val == "1" || val.to_lowercase() == "true" => true,
+        _ => false,
+    };
     let mut gpa = Gpapi::new(locale, timezone, device);
 
     if let Err(err) = gpa.login(username, password).await {
@@ -40,10 +44,13 @@ pub async fn download_apps(
                     if sleep_duration > 0 {
                         sleep(TokioDuration::from_millis(sleep_duration)).await;
                     }
-                    match gpa.download(&app_id, None, Path::new(outpath)).await {
+                    match gpa.download(&app_id, None, split_apk, Path::new(outpath)).await {
                         Ok(_) => println!("{} downloaded successfully!", app_id),
                         Err(err) if matches!(err.kind(), GpapiErrorKind::FileExists) => {
                             println!("File already exists for {}. Skipping...", app_id);
+                        }
+                        Err(err) if matches!(err.kind(), GpapiErrorKind::DirectoryExists) => {
+                            println!("Split APK directory already exists for {}. Skipping...", app_id);
                         }
                         Err(err) if matches!(err.kind(), GpapiErrorKind::InvalidApp) => {
                             println!("Invalid app response for {}. Skipping...", app_id);
@@ -53,11 +60,11 @@ pub async fn download_apps(
                         }
                         Err(_) => {
                             println!("An error has occurred attempting to download {}.  Retry #1...", app_id);
-                            match gpa.download(&app_id, None, Path::new(outpath)).await {
+                            match gpa.download(&app_id, None, split_apk, Path::new(outpath)).await {
                                 Ok(_) => println!("{} downloaded successfully!", app_id),
                                 Err(_) => {
                                     println!("An error has occurred attempting to download {}.  Retry #2...", app_id);
-                                    match gpa.download(&app_id, None, Path::new(outpath)).await {
+                                    match gpa.download(&app_id, None, split_apk, Path::new(outpath)).await {
                                         Ok(_) => println!("{} downloaded successfully!", app_id),
                                         Err(_) => {
                                             println!("An error has occurred attempting to download {}. Skipping...", app_id);